# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.1
paths:
  /tasks:
    get:
      summary: Get all tasks
      tags:
        - tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{id}:
    # get:
    #   summary: Get all tasks of user
    #   tags:
    #     - tasks
    #   parameters:
    #     - name: id
    #       in: path
    #       description: The ID of the task
    #       required: true
    #       schema:
    #         type: integer
    #         format: uint
    #   responses:
    #     '200':
    #       description: A list of tasks
    #       content:
    #         application/json:
    #           schema:
    #             type: array
    #             items:
    #               $ref: '#/components/schemas/Task'
    patch:
      summary: Update a task by ID
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          description: The ID of the task
          required: true
          schema:
            type: integer
            format: uint
            x-go-name: ID  # Переименование поля в Go
      requestBody:
        description: The data task to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: The updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete a task by ID
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          description: The ID of the task
          required: true
          schema:
            type: integer
            format: uint
            x-go-name: ID  # Переименование поля в Go
      responses:
        '204':
          description: Task deleted
  /users:
    get:
      summary: Get all users
      tags:
        - users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      summary: Get all tasks of user
      tags:
        - users
      parameters:
        - name: id
          in: path
          description: The ID of the user
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    patch:
      summary: Update a user by ID
      tags:
        - users
      parameters:
        - name: id
          in: path
          description: The ID of the user
          required: true
          schema:
            type: integer
            format: uint
            x-go-name: ID  # Переименование поля в Go
      requestBody:
        description: The data user to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete a user by ID
      tags:
        - users
      parameters:
        - name: id
          in: path
          description: The ID of the user
          required: true
          schema:
            type: integer
            format: uint
            x-go-name: ID  # Переименование поля в Go
      responses:
        '204':
          description: User deleted
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: uint # так как в БД айди это только положительные числа.
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: uint # так как в БД айди это только положительные числа.
    User:
      type: object
      properties:
        id:
          type: integer
          format: uint # так как в БД айди это только положительные числа.
        email:
          type: string
        password:
          type: string